feat: Finished part 4 for Java course

+++ Java by Helsinki University/mooc-java-programming-i/part03-Part03_32.PersonalDetails/src/main/java/PersonalDetails.java
@@ -32,7 +32,7 @@ public class PersonalDetails {
         for (int birthYear : birthYears) {
             birthYearsSum += birthYear;
         }
-        double birthYearsAverage = birthYearsSum / birthYears.size();
+        double birthYearsAverage = birthYearsSum / (double) birthYears.size();
         
         System.out.println("Longest name: " + longestName);
         System.out.println("Average of the birth years: " + birthYearsAverage);
